TestDataLoader relocation
The TestDataLoader class has been moved to the
appropriate test folder. This minimizes security risks
and allows the code to be sepereted properly.

The plugin for SpringBoot run has been deleted because
it is replaced with a run profile. You now run the
application by selecting IpsenSpineApplication and
then clicking either the run or the debug button next
to it.